fix(flutter): resolve remaining dart logic and test defects in driver/customer apps

+++ apps/customer/lib/widgets/order_card.dart
@@ -182,7 +182,7 @@ class HistoryOrderCard extends StatelessWidget {
                 ),
                 if (order.ratingGiven != null)
                   Text(
-                    '⭐' * order.ratingGiven!,
+                    '⭐' * order.ratingGiven!.toInt(),
                     style: const TextStyle(fontSize: 13),
                   )
                 else
